Move `nls` localization namespace to common
