+++ packages/core/src/common/command.ts
@@ -18,8 +18,7 @@ import { injectable, inject, named } from 'inversify';
 import { Event, Emitter, WaitUntilEvent } from './event';
 import { Disposable, DisposableCollection } from './disposable';
 import { ContributionProvider } from './contribution-provider';
-// eslint-disable-next-line @theia/runtime-import-check
-import { nls } from '../browser/nls';
+import { nls } from './nls';
 
 /**
  * A command is a unique identifier of a function
