+++ packages/core/src/common/index.ts
@@ -38,6 +38,7 @@ export * from './strings';
 export * from './application-error';
 export * from './lsp-types';
 export * from './contribution-filter';
+export * from './nls';
 
 import { environment } from '@theia/application-package/lib/environment';
 export { environment };
